refactor code design
